Trim mdlManager parameter width

+++ src/Event/Fcl/fclShopMisc.c
@@ -269,7 +269,7 @@ u64 FUN_003f33d0(int param_1,u32 param_2);
 u64 FUN_003f3970(long param_1,u32 param_2);
 u64 FUN_003f39a0(long param_1,long param_2,u32 param_3);
 void FUN_003f3dc0(int param_1,u64 param_2);
-u32 FUN_003f3e60(int param_1,u64 param_2);
+u32 FUN_003f3e60(int param_1,u32 param_2);
 u32 FUN_003f3f00(u64 param_1);
 void FUN_003f3fc0(int param_1,int param_2,u32 param_3,u64 param_4);
 void FUN_003f4350(int param_1,int param_2,int param_3,int param_4);
@@ -3989,7 +3989,7 @@ void FUN_003f3dc0(int param_1,u64 param_2)
 // FUN_003F3E60 NONMATCHING
 
 
-u32 FUN_003f3e60(int param_1,u64 param_2)
+u32 FUN_003f3e60(int param_1,u32 param_2)
