+++ src/itfPanel.c
@@ -3584,7 +3584,7 @@ u64 FUN_003ace60(u64 param_1,int param_2)
 u32 FUN_003acf80(u32 param_1,int param_2)
 {
   u8 bVar2;
-  u8 uVar3;
+  u32 uVar3;
   int lVar4;
   u32 packed;
   u8 *pbVar5;
